requirements.txt
tensorflow>=2.15,<2.18
roboflow>=1.1.20
python-dotenv>=1.0.1
ultralytics>=8.3.0
pillow>=10.0.0
opencv-python>=4.8.0
